+++ 곽준선/kjs/week10/src/main/kotlin/kotlinassignment/week10/domain/member/service/MemberService.kt
@@ -1,46 +1,20 @@
 package kotlinassignment.week10.domain.member.service
 
-import kotlinassignment.week10.domain.exception.DuplicatedEmailException
 import kotlinassignment.week10.domain.member.dto.MemberLoginRequest
 import kotlinassignment.week10.domain.member.dto.MemberLoginResponse
 import kotlinassignment.week10.domain.member.dto.MemberSignUpRequest
-import kotlinassignment.week10.domain.member.exception.InvalidCredentialException
-import kotlinassignment.week10.domain.member.model.Member
-import kotlinassignment.week10.domain.member.repository.MemberRepository
-import kotlinassignment.week10.infra.jwt.JwtUtil
-import org.springframework.security.crypto.password.PasswordEncoder
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 
-@Service
-class MemberService(
-    private val memberRepository: MemberRepository,
-    private val passwordEncoder: PasswordEncoder,
-    private val jwtUtil: JwtUtil,
-) {
+interface MemberService {
 
-    fun login(request: MemberLoginRequest): MemberLoginResponse {
-        val user = memberRepository.findByEmail(request.email) ?: throw InvalidCredentialException()
+    /**
+     * @param (type: MemberLoginRequest) 회원 로그인을 위한 DTO
+     * @return (type: MemberLoginResponse) 로그인 후 반환하는 데이터
+     */
+    fun login(request: MemberLoginRequest): MemberLoginResponse
 
-        if (!passwordEncoder.matches(request.password, user.password)) {
-            throw InvalidCredentialException()
-        }
-
-        return MemberLoginResponse(accessToken = jwtUtil.generateAccessToken(user.id.toString(), user.email))
-    }
-
-    @Transactional
-    fun signUp(request: MemberSignUpRequest): Unit {
-        if (memberRepository.existsByEmail(request.email)) {
-            throw DuplicatedEmailException()
-        }
-
-        memberRepository.save(
-            Member(
-                email = request.email,
-                password = passwordEncoder.encode(request.password),
-                nickname = request.nickname,
-            )
-        )
-    }
+    /**
+     * @param (type: MemberSignUpRequest) 회원 가입을 위한 DTO
+     * @return 없음
+     */
+    fun signUp(request: MemberSignUpRequest): Unit
 }
